Refresh sample data; only decompress valid UPX files
