+++ tests/linux/2024.vncjew/__min__c.simple
@@ -1,13 +1,71 @@
-# linux/2024.vncjew/__min__c: critical
-anti-static/elf/entropy: high
-anti-static/elf/header: high
-anti-static/elf/multiple: medium
-anti-static/packer/upx: high
-c2/addr/ip: high
+# linux/2024.vncjew/__min__c ∴ /__min__c: critical
+c2/addr/ip: medium
 c2/addr/url: low
 c2/tool_transfer/arch: low
-fs/proc/self_exe: medium
+credential/password: low
+credential/ssl/private_key: low
+crypto/aes: low
+crypto/cipher: medium
+crypto/decrypt: low
+crypto/ecdsa: low
+crypto/ed25519: low
+crypto/encrypt: medium
+crypto/public_key: low
+crypto/rc4: low
+crypto/tls: low
+data/compression/gzip: low
+data/encoding/base64: low
+data/encoding/json: low
+data/encoding/json_decode: low
+data/hash/md5: low
+discover/system/cpu: low
+discover/system/hostname: low
+discover/system/platform: low
+discover/user/HOME: low
+discover/user/USER: low
+evasion/bypass_security/linux/iptables: medium
+evasion/bypass_security/linux/iptables_append: medium
+exec/plugin: low
+exec/program: medium
+fs/directory/list: low
+fs/file/open: low
+fs/file/read: low
+fs/link_read: low
+fs/path/etc: low
+fs/path/etc_hosts: medium
+fs/path/etc_resolv.conf: low
+fs/path/home: low
+fs/permission/chown: medium
+fs/permission/modify: medium
+fs/tempfile: low
+malware/family/vncjew: critical
+net/dns: low
+net/dns/servers: low
 net/dns/txt: low
+net/http/accept: low
+net/http/accept_encoding: low
+net/http/auth: low
+net/http/cookies: medium
+net/http/post: medium
+net/http/proxy: low
 net/http/request: low
-net/remote_control/vnc: medium
+net/http/websocket: medium
+net/ip/addr: medium
+net/ip/host_port: medium
+net/ip/multicast_send: low
+net/ip/parse: medium
+net/resolve/hostname: low
+net/socket/listen: medium
+net/socket/local_addr: low
+net/socket/peer_address: low
+net/socket/receive: low
+net/socket/send: low
+net/tcp/connect: medium
+net/udp/receive: low
+net/udp/send: low
 net/url/embedded: low
+net/url/parse: low
+net/url/request: medium
+os/fd/sendfile: low
+os/kernel/netlink: low
+sec-tool/net/masscan: high
