+++ tests/linux/clean/trino.linux-amd64.launcher.simple
@@ -1,14 +1,60 @@
-# linux/clean/trino.linux-amd64.launcher: medium
-anti-static/elf/content: medium
-anti-static/elf/entropy: medium
-anti-static/elf/header: medium
-anti-static/elf/multiple: medium
-anti-static/packer/upx: medium
+# linux/clean/trino.linux-amd64.launcher ∴ /trino.linux-amd64: medium
+c2/addr/ip: medium
 c2/addr/url: low
 c2/tool_transfer/arch: low
+collect/archives/zip: medium
+credential/password: low
+credential/ssl/private_key: low
 crypto/aes: low
+crypto/ecdsa: low
+crypto/public_key: low
+crypto/rc4: low
+crypto/tls: low
 data/compression/gzip: low
-fs/proc/self_exe: medium
+data/encoding/base64: low
+discover/system/cpu: low
+discover/system/hostname: low
+discover/system/platform: low
+exec/plugin: low
+exec/program: medium
+fs/directory/create: low
+fs/directory/remove: low
+fs/file/delete: low
+fs/file/open: low
+fs/file/read: low
+fs/link_read: low
+fs/lock_update: low
+fs/path/etc: low
+fs/path/etc_hosts: medium
+fs/path/etc_resolv.conf: low
+fs/path/users: medium
+fs/path/var: low
+fs/permission/chown: medium
+fs/permission/modify: medium
+net/dns: low
+net/dns/servers: low
 net/dns/txt: low
+net/http/auth: low
 net/http/post: medium
+net/http/proxy: low
+net/http/request: low
+net/ip/addr: medium
+net/ip/host_port: medium
+net/ip/parse: medium
+net/resolve/hostname: low
+net/socket/listen: medium
+net/socket/local_addr: low
+net/socket/peer_address: low
+net/socket/receive: low
+net/socket/send: low
+net/tcp/connect: medium
+net/udp/receive: low
+net/udp/send: low
 net/url/embedded: low
+net/url/parse: low
+net/url/request: medium
+os/fd/sendfile: low
+os/kernel/netlink: low
+persist/daemon: medium
+persist/pid_file: medium
+process/groups_set: low
